requirements.txt
asgiref==3.4.1
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.3
coverage==5.5
dj-database-url==0.5.0
Django==3.2.5
django-cors-headers==3.7.0
django-filter==2.4.0
django-heroku==0.3.1
django-modeltranslation==0.17.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
easy-thumbnails==2.7.1
flake8==3.9.2
gunicorn==20.1.0
idna==3.2
iniconfig==1.1.1
Markdown==3.3.4
mccabe==0.6.1
packaging==21.0
Pillow==8.3.0
pluggy==1.0.0
polib==1.1.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.5
pytest-cov==2.12.1
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6
whitenoise==5.2.0
